feat: move Firestore repos and Auth to core:data commonMain, clean up :app
Move FireStoreModels, FireStoreRepository (6 interfaces + 6 impls),
AuthRepository/AuthRepositoryImpl, and Koin bindings to core:data commonMain.
Merge firestoreModule + auth bindings into unified coreDataModule.
Remove dead AppContainer, move UpdateTeamUseCaseImpl to proper package,
fix cross-module smart cast on nullable Float fields, strip auth bindings
from app DataModule.
